backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gunicorn>=21.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
bcrypt>=4.0.0
slowapi>=0.1.9

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# Scraping & Data Ingestion
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.0

# Caching & Background Tasks
redis>=5.0.0
celery>=5.3.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Monitoring & Logging
python-json-logger>=2.0.0
sentry-sdk[fastapi]>=1.40.0

firecrawl-py>=1.0.0